Use `{:packet, 4}` option in TCP server
This establishes a protocol in which the first four bytes sent to or from the `:gen_tcp` server contain the packet length and are followed by that many bytes of packet body. For incoming data, the `ElixirSense.Server.TCPServer` module was doing basically this, but doing it manually. With this change, the server no longer needs to keep track of the "rest" between packets.

More importantly, though, this greatly simplifies handling the response data on the client. In Java, for example, it's helpful to know how big a buffer to allocate for incoming packets.

The Coffeescript client needs to be updated, but it should require only a simple change to omit the prefix byte (with value `101`) when sending requests and to ignore the first four bytes of the response.